+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - AA-129 - wagtail import export for Beta environment
 - GP2-36 - lesson completed functionality
 - no ticket - increase token expiration date to 5 days from 1
+- AA-155 - healthcheck endpoint to wrap CCCE API request
 - GP2-933 - contact us
 
 ## Pre-release